requirements.txt
blinker            1.9.0
certifi            2025.4.26
charset-normalizer 3.4.2
click              8.2.1
colorama           0.4.6
Flask              3.1.1
flask-cors         6.0.0
Flask-Mail         0.10.0
Flask-SQLAlchemy   3.1.1
Flask-WTF          1.2.2
greenlet           3.2.2
idna               3.10
itsdangerous       2.2.0
Jinja2             3.1.6
MarkupSafe         3.0.2
packaging          25.0
pillow             11.2.1
pip                25.0.1
pytesseract        0.3.13
python-dotenv      1.1.0
requests           2.32.3
SQLAlchemy         2.0.41
tesseract          0.1.3
typing_extensions  4.13.2
urllib3            2.4.0
Werkzeug           3.1.3
WTForms            3.2.1